Setup extension boilerplate (#2)
* added vitest support to the extension

* study log extension directory name change

* vitest to jest switch;
jest dependencies and config setup

* add jest commands to package.json

* Run prettier on webextension code

* Tweak `README.md` formatting

* Change `package.json` to be consistent with our other packages

* Remove redundant `jest` prefix from test scripts

* Add plain `test` script for turbo

* Add turbo config for tests

* Remove stray `package-lock.json`

* Add shortcut scripts for running tests

* Update `pnpm-lock.yaml`

* Use lockfile from main branch to avoid conflict
