--- apps/studylog-extension/src/entries/background/main.ts
@@ -1,5 +1,0 @@
-import browser from "webextension-polyfill";
-
-browser.runtime.onInstalled.addListener(() => {
-  console.log("Extension installed");
-});
