--- apps/studylog-extension/src/entries/background/script.ts
@@ -1,1 +1,0 @@
-import "./main";
