--- apps/studylog-extension/src/entries/background/serviceWorker.ts
@@ -1,1 +1,0 @@
-import "./main";
