--- apps/studylog-extension/src/entries/popup/main.ts
@@ -1,4 +1,0 @@
-import { createApp } from "vue";
-import App from "./App.vue";
-
-createApp(App).mount("#app");
