+++ apps/web-extension/README.md
@@ -6,7 +6,7 @@ This template should help get you started developing a vue web extension in Vite
 
 The extension manifest is defined in `src/manifest.js` and used by `@samrum/vite-plugin-web-extension` in the vite config.
 
-Background, content scripts, options, and popup entry points exist in the `src/entries` directory. 
+Background, content scripts, options, and popup entry points exist in the `src/entries` directory.
 
 Content scripts are rendered by `src/entries/contentScript/renderContent.js` which renders content within a ShadowRoot
 and handles style injection for HMR and build modes.
@@ -26,40 +26,46 @@ See [Vite Configuration Reference](https://vitejs.dev/config/).
 
 ## Project Setup
 
-```sh
-npm install
+```console
+$ pnpm install
 ```
 
 ## Commands
+
 ### Build
+
 #### Development, HMR
 
 Hot Module Reloading is used to load changes inline without requiring extension rebuilds and extension/page reloads
-```sh
-npm run dev
+
+```console
+$ pnpm dev
 ```
 
 #### Development, Watch
 
 Rebuilds extension on file changes. Requires a reload of the extension (and page reload if using content scripts)
-```sh
-npm run watch
+
+```console
+$ pnpm watch
 ```
 
 #### Production
 
 Minifies and optimizes extension build
-```sh
-npm run build
+
+```console
+$ pnpm build
 ```
 
 ### Load extension in browser
 
 Loads the contents of the dist directory into the specified browser
-```sh
-npm run serve:chrome
+
+```console
+$ pnpm serve:chrome
 ```
 
-```sh
-npm run serve:firefox
+```console
+$ pnpm serve:firefox
 ```
