+++ apps/web-extension/package.json
@@ -1,29 +1,36 @@
 {
-  "name": "vite-web-extension",
-  "version": "1.0.0",
-  "displayName": "Vite Web Extension",
-  "author": "@samrum/vite-plugin-web-extension",
+  "name": "@studylog/web-extension",
+  "private": true,
+  "version": "0.0.0",
+  "displayName": "StudyLog",
   "description": "A @samrum/vite-plugin-web-extension web extension",
   "scripts": {
     "build": "vue-tsc --noEmit && vite build",
     "watch": "vite build --watch",
     "dev": "vite",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage=true",
     "serve:firefox": "web-ext run --start-url \"about:debugging#/runtime/this-firefox\"  --source-dir ./dist/",
     "serve:chrome": "web-ext run -t chromium --start-url \"https://google.com\" --source-dir ./dist/"
   },
   "license": "MIT",
   "devDependencies": {
+    "@babel/core": "^7.20.5",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-typescript": "^7.18.6",
     "@samrum/vite-plugin-web-extension": "^2.0.0",
     "@types/chrome": "^0.0.196",
-    "@types/webextension-polyfill": "^0.9.0",
     "@vitejs/plugin-vue": "^3.1.0",
+    "@vitest/coverage-c8": "^0.25.8",
+    "babel-jest": "^29.3.1",
+    "mockzilla-webextension": "^0.15.0",
     "typescript": "^4.8.2",
     "vite": "~3.1.0",
     "vue-tsc": "^0.40.8",
     "web-ext": "^7.2.0"
   },
   "dependencies": {
-    "vue": "^3.2.38",
-    "webextension-polyfill": "^0.10.0"
+    "vue": "^3.2.38"
   }
 }
