+++ apps/web-extension/src/entries/background/main.ts
@@ -1,0 +1,5 @@
+import browser from 'webextension-polyfill';
+
+browser.runtime.onInstalled.addListener(() => {
+  console.log('Extension installed');
+});
