+++ apps/web-extension/src/entries/background/script.ts
@@ -1,0 +1,1 @@
+import './main';
