+++ apps/web-extension/src/entries/background/serviceWorker.ts
@@ -1,0 +1,1 @@
+import './main';
