+++ apps/web-extension/src/entries/background/test/example.test.ts
@@ -1,0 +1,5 @@
+// jest test example
+
+it('test example', () => {
+  expect(Math.sqrt(4)).toBe(2);
+});
