+++ apps/web-extension/src/entries/contentScript/primary/App.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script setup lang="ts">
-import logo from "~/assets/logo.svg";
+import logo from '~/assets/logo.svg';
 
 const logoImageUrl = new URL(logo, import.meta.url).href;
 </script>
