+++ apps/web-extension/src/entries/options/App.vue
@@ -1,11 +1,11 @@
 <template>
   <main>
-    <PageContent> Popup </PageContent>
+    <PageContent>Options</PageContent>
   </main>
 </template>
 
 <script setup lang="ts">
-import PageContent from "~/components/PageContent.vue";
+import PageContent from '~/components/PageContent.vue';
 </script>
 
 <style>
