+++ apps/web-extension/src/entries/popup/App.vue
@@ -1,11 +1,11 @@
 <template>
   <main>
-    <PageContent> Options </PageContent>
+    <PageContent>Popup</PageContent>
   </main>
 </template>
 
 <script setup lang="ts">
-import PageContent from "~/components/PageContent.vue";
+import PageContent from '~/components/PageContent.vue';
 </script>
 
 <style>
