+++ apps/web-extension/src/env.d.ts
@@ -1,8 +1,8 @@
 /// <reference types="vite/client" />
 /// <reference types="@samrum/vite-plugin-web-extension/client" />
 
-declare module "*.vue" {
-  import type { DefineComponent } from "vue";
+declare module '*.vue' {
+  import type { DefineComponent } from 'vue';
   // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/ban-types
   const component: DefineComponent<{}, {}, any>;
   export default component;
