+++ package.json
@@ -6,6 +6,8 @@
     "build": "turbo run build",
     "dev": "turbo run dev",
     "lint": "turbo run lint",
+    "test": "turbo run test",
+    "test:watch": "turbo run test:watch",
     "format": "prettier --write \"**/*.{{c,m,}js,{c,m,}ts,md,vue,json,yaml,yml}\""
   },
   "engines": {
